// src/pages/Homepage/models/movie.ts
export type Movie = {
  Title: string;
  Year: string;
  Rated: string;
  Released: string;
  Runtime: string;
  Genre: string;
  Director: string;
  Writer: string;
  Actors: string;
  Plot: string;
  Language: string;
  Country: string;
  Awards: string;
  Poster: string;
  Metascore: string;
  Rating: string;
  Votes: string;
  Id: string;
  Type: string;
  DVD: string;
  BoxOffice: string;
  Production: string;
};